more fixes for SRIOV detection
This has a really ugly sleep(5) before the PCI bus scan, to allow the
PCI bus to settle down after creating 126 new devices just before udev
is invoked on each of those.  Without a delay, we can't know the port
number on the card, as the PF may not get created before the VF (and
in fact is observed to invoke udev only _after_ udev is invoked for
the VFs).

This of course introduces delay into the system startup, at least 5
seconds, and there's no guarantee that 5 seconds is enough if a bunch
of other things are happening at the same time.  Welcome to the nature
of udev...