copy SMBIOS slot info to all matching domain/bus/device/functions
Most of the time, there will be only a single SMBIOS Type 9 slot entry
for the "base" function (function 0) on the device in that slot.
But, if we have a multi-function card, we want each of the functions to
have that same information.  So, save the SMBIOS information if it is
an exact match to a PCI d/b/d/f, and fill out other functions too, but
if there is already an exact match for a sub-function, leave it alone.