Don't assign names to unknown devices
Just because we know the names of some devices, doesn't mean we know
the names of all.  In the case shown here:
  https://bugzilla.redhat.com/show_bug.cgi?id=662905
we know the name of the add-in card, but should not know the name of
the embedded NIC because it's not represented in SMBIOS or PIRQ. Don't
assign a name we don't know to be good.