start with port index 1, not index 0
Dell PowerEdge Servers name their embedded NICs starting at 1, as seen
in the SMBIOS tables.  HP servers do likewise in their SMBIOS entries.

For consistency on machines with only 1 embedded NIC, and which don't have
entries in SMBIOS to provide a different index (so we fall back to
PIRQ), we want that named em1 to match servers above, not em0 which
wouldn't match.  This way, users can use the name em1 regardless of if
the system provides the info in SMBIOS or not.  Of course, SMBIOS may
provide a different number, in which case that is honored.

This means you will see:
  em<port>  where port starts at 1
  pci<slot>#<port> where port starts at 1