properly look for SMBIOS, then $PIR, then recurse
We were looking at SMBIOS and recursing up the tree looking for parent
slot information, and only if that failed were we starting from the
bottom again with $PIR, and then recursing up.  This was causing
incorrect lookup when a device appears natively in PIR in a slot but it's parent
had already been set as embedded - both would appear embedded which
was incorrect.