From Adrian Huang
The command "biosdevname -d" shows the field "SMBIOS Lable" with "MEZZ2" for
all network interfaces when there are two or more PCI root ports in the same
Bus ID/Device ID.

With this patch, the info can be showed correctly if there are two or more PCI
root ports in the same Bus ID/Device ID. For example, the field of the
"p4p1" and "p4p2" would be showed as "MEZZ1".

This patch is based on v0.5.1.

Signed-off-by: Adrian Huang <ahuang12@lenovo.com>